Bodge a fix for crash on exit with voxel tools
I am not sure if the amount of time to wait for depends on the world size, objects being unloaded, etc. If it does, then I may find another workaround because I do not know enough about voxel tools to patch a bug related specifically to its C# bindings. The crash on exit is specifically related to a C# script_bindings container not being empty upon unloading.